AbstractFunctionRestrictionsSniff::is_targetted_token(): prevent false positives on PHP 8.0+ attributes
`T_STRING` tokens in PHP 8.0+ attributes are either class names or possibly constant names (as a parameter for the class instantiation).
They are never function calls (as things are).

This commit ensures that `T_STRING` tokens in attributes are not confused with function calls.

Tested via tests in the test file for the `WordPress.WP.DiscouragedFunctions` sniff.