Move `is_class_object_call()` utility method to dedicated `ContextHelper`
The `is_class_object_call()` utility method is only used by a small set of sniffs, so is better placed in a dedicated class.

This commit moves the `is_class_object_call()` method to the new `WordPressCS\WordPress\Helpers\ContextHelper` class and starts using that class in the relevant sniffs.

The method has also been renamed to `has_object_operator_before()` as it doesn't actually check if something is a (method) _call_, just that something OO (method, constant, property) is being accessed.

Related to 1465

This method is tested via the `WordPress.WP.DiscouragedFunctions` sniff.