Move `is_token_namespaced()` utility method to dedicated `ContextHelper`
The `is_token_namespaced()` utility method is only used by a small set of sniffs, so is better placed in a dedicated class.

This commit moves the `is_token_namespaced()` method to the new `WordPressCS\WordPress\Helpers\ContextHelper` class and starts using that class in the relevant sniffs.

Related to 1465

This method is tested via the `WordPress.WP.DiscouragedFunctions` sniff.