Move `is_in_isset_or_empty()` utility method to dedicated `ContextHelper`
The `is_in_isset_or_empty()` utility method is only used by a small set of sniffs, so is better placed in a dedicated class.

This commit moves the `is_in_isset_or_empty()` method to the new `WordPressCS\WordPress\Helpers\ContextHelper` class and starts using that class in the relevant sniffs.

Related to 1465

This method will be tested via the `WordPress.Security.NonceVerification` sniff (via pre-existing and new tests).