AbstractFunctionRestrictionsSniff: add test with empty functions group
This test is added to the `WordPress.DB.RestrictedFunctions` test class as we need to test the sniff with both the `$unittest_groups` property set and not set, so we need to use a second class to test the abstract properly.

The `WordPress.DB.RestrictedFunctions` sniff was chosen deliberately as it is one of the few function related sniff which use the `'allow'` key, which will allow us to test the correct handling of that key as well.