AbstractClassRestrictionsSniff::is_targetted_token(): bug fix - sniff does not bow out early on hierarchy keywords
The abstract class turns every found name into a fully qualified name, so, as things are, the hierarchy keywords would be prefixed with a `\` too, so the `in_array()` comparison would never match.

While this doesn't lead to false positives, it should still be fixed for performance reasons.

Fixed now.

Includes unit test.