AbstractArrayAssignmentRestrictions: remove support for 'callback'
[OPTIONAL]

The `getGroups()` method allows to pass a custom callback for any particular group, however, in practice that array key is never used as the `callback()` method (the default) is declared as `abstract` and therefore **must** be defined in a child class.

I propose to remove support for the `callback` array key from the abstract sniff.

Sniffs which extend this abstract can still use different logic for different groups by applying a switch to the `$group` parameter received in the `callback()`, along the lines of:
```php
public function callback( $key, $val, $line, $group ) {
    switch( $group ) {
        case 'groupA':
            return $this->callback_for_group_A($key, val, $line);
        case 'groupB':
            return $this->callback_for_group_B($key, val, $line);
    }
}
```

I have verified that [none of the sniffs published in public repos and which extend this abstract sniff](https://sourcegraph.com/search?q=context:global+use+WordPressCS%5CWordPress%5CAbstractArrayAssignmentRestrictionsSniff&patternType=standard&case=yes&sm=0&groupBy=group) currently use the `'callback'` key.

Removing the support for the key is a breaking change, but one with low impact based on current usage of the abstract.