PrintingFunctionsTrait: make sure function names are checked case-insensitively
These functions should be self-contained, so should not presume that the sniff has already lowercased the function name before passing it.

This fixes a bug as, in this case, the sniff didn't actually lowercase the name before passing it to the trait method, so the sniff would throw false negatives for non-lowercase function calls.

Tested by adjusting some pre-existing tests for the `EscapeOutput` sniff.