Files/FileName: improve handling of files using non-underscore, non-dash word separators
Addresses Gary's question in 2154:

> I can see there are some test files with extra `.` in (and not just underscores), but I was wondering if an incorrect file name like `class.my-class.php` might be detected, as the checking for hyphenation only replaces underscores.

I have tested this and while the file would already be flagged by the "class name vs file name check", the "is hyphenated" check would not flag the file.

Fixed now. Includes extra tests.

Note: this fix requires that the numbered test files we were using have to be renamed as the `.#.inc` would cause the sniff to throw errors when there shouldn't be any.

Related to issue 2164, but doesn't fix it completely.