WP/DeprecatedFunctions: update the functions list based on WP 6.3-RC1
Updates the list of deprecated functions in the `WP.DeprecatedFunctions` sniff based on the WP 6.3.0-RC1 release.
Input for this based on @JDGrimes's WP deprecated code scanner, though that is getting more and more difficult to run as it is no longer maintained and is throwing lots of errors.

Previous: 2121