Files/FileName: update for changes in WP Core
In WP 6.1.0, the classes contained in the files listed in `$class_exceptions` were moved to files which comply with the file name requirements.
The original files still remain, but no longer contain a class, so we no longer need to make an exception for these in the `check_filename_has_class_prefix()` check.

However, with the stricter check for hyphenation-only word separators as introduced in the earlier commit, we do need to continue making an exception for these files for the `check_filename_is_hyphenated()` check.

For that reason, the properties remain.

Having said that, in this commit:
* I remove the check against these properties from the `check_filename_has_class_prefix()` check.
* The properties are renamed to `$*hyphenation_exceptions` to better reflect what they are used for.
* Two additional non-compliant files from WP Core are added to the list.