Security/NonceVerification: efficiency fix - check for being in a nonce verification method early
As things were, if the superglobal being examined was _in_ a call to one of the nonce verification functions, the sniff would still loop from the start of the scope (start of the function/file) to the current token before coming to that conclusion.
This is, of course, highly inefficient for long functions/large files.

I've been in two minds about whether to add this check to the `needs_nonce_check()` method or the `has_nonce_check()` method.
For optimal efficiency, the best place is at the start of the `needs_nonce_check()` method as it will prevent having to do all the other "is in type check, is in sanitization function" checks, which are redundant if the call is a nonce check itself.
And after all, if the current superglobal is _in_ a nonce check, it doesn't _need_ a nonce check as it already has one.

To allow for caching the found nonce check, the gathering of the basic information needed for the caching has to be moved to the `process_token()` method though, so both the `needs_nonce_check()` method, as well as the `has_nonce_check()` method now get an extra `$cache_keys` parameter.

It also means that the lists of custom and known nonce verification functions have to be merged earlier, so the call to `$this->mergeFunctionLists()` has been moved up.