Security/NonceVerification: test with empty()
There were plenty of tests which checked the nonce verification in combination with a superglobal being used in `isset()` and `array_key_exists()`, but none in combination with `empty()`, while the sniff does handle that as well and treats it the same as `isset()`/`array_key_exists()`.

This adjusts a few pre-existing tests to make sure that the handling of `empty()` is covered by tests too.

Note: it is important that these have separate tests as `empty()` has its own token.