Arrays/ArrayDeclarationSpacing: implement PHPCSUtils  [2] / PHP 7.2+ keyed lists
Since PHP 7.2, lists can also contain keys and the double arrow in those will tokenize as `T_DOUBLE_ARROW`, same as for arrays.
An array value can contain a list assignment, so we need to take this into account.

The PHPCSUtils `Arrays::getDoubleArrowPtr()` method does so correctly since PHPCSUtils 1.0.7, so upping the minimum supported PHPCSUtils version in combination with switching to the PHPCSUtils method for retrieving the double arrow will fix this.

Includes removing the "initial check" for a double arrow as that only serves to cause the array to be walked twice when an arrow is found, so it's more efficient to not do this initial check.

Includes tests safeguarding the issue with lists.

:point_right: Reviewing with _ignore whitespace changes_ recommended.