Arrays/ArrayDeclarationSpacing: replace whitespace checks with PHPCSExtra sniff
This commit removes the brace related spacing checks from the `WordPress.Arrays.ArrayDeclarationSpacing` sniff in favour of using the PHPCSExtra `NormalizedArrays.Arrays.ArrayBraceSpacing` sniff.

Includes removing tests which were only testing the brace spacing related issue.
Includes minor updates to the `fixed` file for the remaining tests to account for a fix which will no longer be made by this sniff.

Note: in effect, the combined sniffs will do the same as before, though a few more notices will be thrown.
This is due to the original set-up not throwing errors for "space after opener" and "space before closer", when a "single line array with associative keys must be multi-line" error would be thrown, as the fix would be the same.
These "space after opener" and "space before closer" errors will now show together with the "single line must be multi-line" error and the fixers of the combined sniffs will sort it out between them.

Refs:
* PHPCSStandards/PHPCSExtra 12
* https://github.com/PHPCSStandards/PHPCSExtra/pulls?q=is%3Apr+arraybracespacing+is%3Aclosed