DB/PreparedSQLPlaceholders: prevent token walking too far
When an `implode()` function call is found in a `$query`, the sniff tries to find the preceding text string token to check for quotes around the placeholders.

While it is unlikely that an `implode()` wouldn't be preceded by a text string token + concatenation token, it could still be possible.

This updates the logic to do a negative search for the previous non-empty, non-concat token instead to prevent the sniff from walking too far.

I've not included a test as I couldn't come up with a valid code snippet which would hit a false positive, but it should still be fixed IMO.

:point_right: This commit will be easier to review while ignoring whitespace changes.