Security/EscapeOutput: bug fix - false positives on non-global function calls
As things were, the sniff would look for `T_STRING` tokens with certain contents, but never do a proper check if those were actually function calls to the WP/PHP native _global_ function.

This would lead to false positives for calls to methods, function declarations etc.

By changing the parent class to the `AbstractFunctionRestrictionsSniff`, we can re-use the "if this a global function call" logic already in place in that abstract, while still maintaining the handling of the other tokens in the pre-existing manner.

There are three caveats to this:
1. The `AbstractFunctionRestrictionsSniff` contains an `exclude` property, which allows to silence notices about certain function for a complete group.
    As this is a security sniff, we don't want end-users to be able to silence notices about groups of functions in that way.
    To that end, the ability for end-users to set the `exclude` property for this sniff is being disabled.
2. The `AbstractFunctionRestrictionsSniff` class will call the `process_matched_token()` method once for each match in each function group.
    As the "unsafe printing functions" are also included in the "printing functions", this would lead to the `process_matched_token()` method being called twice for those methods, leading to duplicate error notices.
    I'm preventing this by removing the "unsafe printing functions" from the "printing functions" list prior to passing it off to the parent class.
3. Lastly, the `AbstractFunctionRestrictionsSniff` class expects the list of target functions to be available when the sniff is instantiated. However, in test situations, that list can fluctuate, as the "printing functions" supports the ability to add "custom printing functions", which is done on the fly in the test case file using `// phpcs:set ...`.
    To get round that, I'm re-registering the "function groups" every time a `T_STRING` (potential function call) triggers the sniff, but only when the tests are being run.
    While it is possible for end-users to use the `// phpcs:set` syntax, it is little known and is not mentioned anywhere in the documentation as a feature for end-users, so just re-registering the function groups for test runs would seem sufficient.

Includes tests for situations fixed by this particular commit.
Includes a test using the PHP 8.0+ nullsafe object operator.

Note: the `PrintingFunctionsTrait::is_printing_function()` has now become unused and _could_ be removed.
For consistency with other traits/helper classes, I'm electing to leave it be (for now) though.