Security/EscapeOutput: bug fix - false positives on non-function calls
As things were, the sniff would look for `T_STRING` tokens with certain contents, but never do a proper check if those were actually function calls.

This would lead to false positives, but also potentially to an `Internal.Exception` for the `$end_of_statement` not being defined.

This commit is a first step towards fixing this.

Includes tests for select situations fixed by this particular commit.