Security/EscapeOutput: bug fix - improve start/end determination for print statements
`print` is a language construct and can be used as part of an expression.

The sniff did not take this into account correctly. This would lead to the sniff walking too far and examining code which is not part of the `print` statement. This would lead to both false positives as well as false negatives.

This commit improves that situation, but doesn't completely fix it.

First off, the sniff now uses the `BCFile::findEndOfStatement()` method to more accurately retrieve the end token for the expression. This should improve the handling of the most common cases.
This part includes live coding/parse error protection for unfinished `print` statements at the end of a file.

However, the `[BC]File::findEndOfStatement()` method does not take ternaries into account when determining the end of the expression.

So, secondly, code has been added to special case `print` statements in the "then" part of a ternary and determine the end of such an expression correctly.

Includes tests for situations fixed by this particular commit.

Additional notes:
* I can still come up with more (exotic) test cases with `print` in ternaries which would still not be handled correctly.
    I'm leaving those as "future scope".
    Example:
    ```php
    ( $fop === 1 ) ? true && print $foo : print $bar; // The first print will now examine everything to the end of the statement, not to the end of the expression, which means that `$bar` will not be flagged as the sniff bow out before it reached $bar.
    ```

Fixes 2209