EscapingFunctionsTrait: add wp_readonly() to the list of auto-escaped functions
This function was added in WP 5.9 to replace the `readonly()` function as `readonly` became a reserved keyword in PHP 8.1 (though only partially to allow the WP community some time to catch up).

I'm leaving `readonly()` in the list. While it shouldn't be used anymore, if it is, it still is a valid auto-escaped function.

Ref: https://core.trac.wordpress.org/ticket/53858