Security/EscapeOutput: improve handling of inline expressions
When the sniff would encountered one of the keywords the sniff listens to as an inline expression in the statement which was being examined, the sniff would flag the keyword as needing escaping.

This is wrong as that expression should be examined separately by the sniff (but wouldn't be as the sniff skips to the end of the examined code).

This commit fixes that by adjusting the logic in the loop in the `check_code_is_escaped()` method to check for keywords the sniff listens too and then recursing into the sniff again to examine that expression separately, after which the loop for the original statement will continue from the end of the examined expression.

Includes tests.

Fixes 1861