Update funding information
The WP PHP Open Collective, aside from the name conveying the completely wrong message, has so far not been in touch with us about how funds can be released or anything else.

This seems like a dead-end.

Switching to the PHP_CodeSniffer Open Collective makes more sense (to me). Happy to have a discussion about this.